Updated function "add" under ui.py to "add_to_cart" for greater clarity.
Also updated this naming in the buttons.py file to reflect the new name so the chili button command is now calling the correct function

Updated the timeout time for the WebDriverWait method within the add_to_cart function from 10 seconds to 2 seconds.  This saves a significant portion of time when encountering products that have already been added to the cart.